requirements.txt

streamlit==1.25.0
pandas==2.1.1
sentence-transformers==2.2.2
transformers==4.35.0
pinecone-client[grpc]==2.3.1
pdfplumber==0.10.2
pytesseract==0.3.10
pdf2image==1.16.3
